+++ tests/acceptance_tests/test_backtest.py
@@ -885,13 +885,13 @@ class TestBacktestAcceptanceTestsMarketMaking:
         self.engine.run()
 
         # Assert
-        assert self.engine.kernel.msgbus.sent_count == 23_688
-        assert self.engine.kernel.msgbus.pub_count == 26_806
+        assert self.engine.kernel.msgbus.sent_count == 23_718
+        assert self.engine.kernel.msgbus.pub_count == 26_829
         assert self.engine.iteration == 8_198
         account = self.engine.portfolio.account(self.venue)
         assert account is not None
-        assert account.event_count == 3_530
-        assert account.balance_total(GBP) == Money(-19_351.96, GBP)
+        assert account.event_count == 3_543
+        assert account.balance_total(GBP) == Money(-19_355.51, GBP)
 
 
 @pytest.mark.xdist_group(name="databento_catalog")